+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## 1.1.2
 
-- bug in the entity conponent
+- bug in the entity component
 - layout of the buttons
 
 ## 1.1.1